+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 2.0.0
 - Consider `rule` nodes only when when building ancestor selectors
+- Use `PostCSS` 6
+- Restrict support to `node.js` >= 4
+- Remove `object-assign` dependency
 
 ## 1.0.0
 - Solve complex nesting scenarios scenarios externalizing parent selectors resolution to [postcss-resolve-nested-selector](https://github.com/davidtheclark/)